Move to XPUFunctions
