+++ c10/xpu/XPUFunctions.cpp
@@ -1,5 +1,7 @@
+#include <c10/core/impl/GPUTrace.h>
 #include <c10/util/Exception.h>
 #include <c10/xpu/XPUFunctions.h>
+#include <c10/xpu/XPUStream.h>
 
 #include <vector>
 
@@ -288,6 +290,48 @@ c10::DeviceIndex exchange_device(c10::DeviceIndex to_device) {
 
 c10::DeviceIndex maybe_exchange_device(c10::DeviceIndex to_device) {
   return exchange_device(to_device);
+}
+
+/*
+ * Note [Synchronize Streams on Device]
+ *
+ * device_synchronize waits for all work previously submitted to the SYCL
+ * queues we manage on `device`. Two paths exist:
+ *  1. Fast path (SYCL >= 2026.1 and device exposes `ext_oneapi_device_wait`):
+ *     issue a single `ext_oneapi_wait_and_throw()` -- a true device-wide wait.
+ *     SYCL < 2026.1 is excluded because earlier runtimes could crash (release
+ *     an invalid queue) when a device-wide wait interleaves with XPUGraph
+ *     capture; see https://github.com/pytorch/pytorch/issues/187277.
+ *  2. Legacy path (otherwise): fall back to syncStreamsOnDevice, which only
+ *     drains the SYCL queues we own.
+ */
+void device_synchronize(DeviceIndex device) {
+  if (device == -1) {
+    device = c10::xpu::current_device();
+  }
+  check_device_index(device);
+
+  // TODO: drop the legacy fallback below once a driver supporting
+  // `ext_oneapi_device_wait` is widely deployed across all supported
+  // platforms.
+#if SYCL_COMPILER_VERSION >= 20260100
+  bool ext_oneapi_device_wait_supported = c10::xpu::get_raw_device(device).has(
+      sycl::aspect::ext_oneapi_device_wait);
+#else
+  bool ext_oneapi_device_wait_supported = false;
+#endif
+
+  if (ext_oneapi_device_wait_supported) {
+    c10::xpu::get_raw_device(device).ext_oneapi_wait_and_throw();
+    const c10::impl::PyInterpreter* interp = c10::impl::GPUTrace::get_trace();
+    if (C10_UNLIKELY(interp)) {
+      (*interp)->trace_gpu_device_synchronization(c10::kXPU);
+    }
+  } else {
+    C10_DIAGNOSTIC_PUSH_AND_IGNORED_IF_DEFINED("-Wdeprecated-declarations")
+    syncStreamsOnDevice(device); // also fires the GPUTrace hook
+    C10_DIAGNOSTIC_POP()
+  }
 }
 
 } // namespace c10::xpu
