+++ c10/xpu/XPUStream.cpp
@@ -361,50 +361,25 @@ std::ostream& operator<<(std::ostream& stream, const XPUStream& s) {
 /*
  * Note [Synchronize Streams on Device]
  *
- * device_synchronize waits for all work previously submitted to the SYCL
- * queues we manage on `device`. Two paths exist:
- *  1. Fast path (SYCL >= 2026.1 and device exposes `ext_oneapi_device_wait`):
- *     issue a single `ext_oneapi_wait_and_throw()` -- a true device-wide wait.
- *     SYCL < 2026.1 is excluded because earlier runtimes could crash (release
- *     an invalid queue) when a device-wide wait interleaves with XPUGraph
- *     capture; see https://github.com/pytorch/pytorch/issues/187277.
- *  2. Legacy path (otherwise): walk every reserved queue in each priority
- *     pool and `wait()` on it. This only drains queues we own; SYCL queues
- *     outside our pools are unaffected.
- *
- * syncStreamsOnDevice is a deprecated alias kept for source compatibility;
- * new callers should use device_synchronize directly.
+ * syncStreamsOnDevice waits for all SYCL queues reserved in the stream pools
+ * on `device` to complete, then fires the GPUTrace device-sync hook. It is
+ * deprecated in favor of c10::xpu::device_synchronize
+ * (c10/xpu/XPUFunctions.cpp), which issues a true device-wide wait when the
+ * SYCL runtime supports it and falls back to this function otherwise.
  */
-
-// Note: The stream pools are lazily initialized on the legacy path; the fast
-// path bypasses our pools entirely.
-void device_synchronize(DeviceIndex device) {
+void syncStreamsOnDevice(DeviceIndex device) {
   if (device == -1) {
     device = c10::xpu::current_device();
   }
   check_device_index(device);
 
-  // TODO: drop the legacy fallback below once a driver supporting
-  // `ext_oneapi_device_wait` is widely deployed across all supported
-  // platforms.
-#if SYCL_COMPILER_VERSION >= 20260100
-  bool ext_oneapi_device_wait_supported = c10::xpu::get_raw_device(device).has(
-      sycl::aspect::ext_oneapi_device_wait);
-#else
-  bool ext_oneapi_device_wait_supported = false;
-#endif
-
-  if (ext_oneapi_device_wait_supported) {
-    c10::xpu::get_raw_device(device).ext_oneapi_wait_and_throw();
-  } else {
-    initXPUStreamsOnce();
-    // Initializes the stream pools (once)
-    initDeviceStreamOnce(device);
-    // kStreamsPerPool (32) reserved queues per priority for each device.
-    for (const auto p : c10::irange(max_compile_time_stream_priorities)) {
-      for (const auto i : c10::irange(kStreamsPerPool)) {
-        streams[device][p][i]->wait();
-      }
+  initXPUStreamsOnce();
+  // Initializes the stream pools (once)
+  initDeviceStreamOnce(device);
+  // kStreamsPerPool (32) reserved queues per priority for each device.
+  for (const auto p : c10::irange(max_compile_time_stream_priorities)) {
+    for (const auto i : c10::irange(kStreamsPerPool)) {
+      streams[device][p][i]->wait();
     }
   }
 
@@ -412,10 +387,6 @@ void device_synchronize(DeviceIndex device) {
   if (C10_UNLIKELY(interp)) {
     (*interp)->trace_gpu_device_synchronization(c10::kXPU);
   }
-}
-
-void syncStreamsOnDevice(DeviceIndex device) {
-  device_synchronize(device);
 }
 
 } // namespace c10::xpu
